Updated to Minecraft 1.21.3 and fixed some bugs

+++ src/main/java/de/farbrisus/zoomi/mixin/GameOptionAccessor.java
@@ -15,11 +15,12 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 public class GameOptionAccessor {
 
     @Inject(method = "getFov", at = @At("RETURN"), cancellable = true)
-    public void getZoomLevel(Camera camera, float tickDelta, boolean changingFov, CallbackInfoReturnable<Double> callbackInfo) {
+    public void getZoomLevel(Camera camera, float tickDelta, boolean changingFov, CallbackInfoReturnable<Float> callbackInfo) {
         if (ZoomiClient.isZooming()) {
-            double fov = callbackInfo.getReturnValue();
-            callbackInfo.setReturnValue(fov * ZoomiClient.getZoomLevel());
+            Float fov = callbackInfo.getReturnValue();
+            callbackInfo.setReturnValue(fov * (float) ZoomiClient.getZoomLevel());
         }
         ZoomiClient.manageSmoothCamera();
     }
+
 }
